change return type of SQL-types to int[] add single transaction with GdxSqlitePreparedStatement.commit() enable utf8 test

+++ GdxSqlite/src/de/longri/gdx/sqlite/GdxSqliteCursor.java
@@ -308,13 +308,13 @@ public class GdxSqliteCursor {
     }
 
 
-    void addRow(String[] columnNames, Object[] values, Long[] types) {
+    void addRow(String[] columnNames, Object[] values, int[] types) {
 
         if (names == null) {
             this.names = columnNames;
             this.types = new int[types.length];
             for (int i = 0; i < types.length; i++) {
-                this.types[i] = types[i].intValue();
+                this.types[i] = types[i];
             }
         }
