+++ GdxSqlite/tests/de/longri/gdx/sqlite/GdxSqlitePreparedStatementTest.java
@@ -107,7 +107,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 double val = Double.valueOf(num + "." + num);
@@ -172,7 +172,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 long val = Long.MAX_VALUE - id;
@@ -245,7 +245,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 byte[] val = new byte[]{(byte) id, (byte) id, (byte) id, (byte) id};
@@ -318,7 +318,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 boolean val = id % 2 == 0;
@@ -391,7 +391,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 short val = (short) id;
@@ -464,7 +464,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 float val = Float.valueOf(num + "." + num);
@@ -537,7 +537,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 double val = Double.valueOf(num + "." + num);
@@ -631,7 +631,7 @@ public class GdxSqlitePreparedStatementTest {
         final AtomicInteger cnt = new AtomicInteger(-1);
         db.rawQuery("SELECT * FROM test", new GdxSqlite.RowCallback() {
             @Override
-            public void newRow(String[] columnName, Object[] value, Long[] types) {
+            public void newRow(String[] columnName, Object[] value, int[] types) {
                 int id = cnt.incrementAndGet();
                 String num = Integer.toString(id);
                 short val = (short) id;
